Add color to index.css
