+++ src/index.css
@@ -1,3 +1,6 @@
+:root {
+  --ui-color-primary: rgba(244, 245, 250, 1)
+}
 html, body {
   margin: 0;
   padding: 0;
