+++ src/main.tsx
@@ -1,14 +1,25 @@
 import React from 'react'
-import ReactDOM from 'react-dom/client'
+
+import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
-import { DashboardPage } from './pages/dashboard/dashboard.page.tsx'
 
 import './index.css'
+import {
+  MainLayout,
+  MenuContainer
+} from './pages/dashboard/layout/main/main.layout.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <BrowserRouter>
-      <DashboardPage />
+      <div style={{ display: 'flex' }}>
+        <MenuContainer>
+          <p>menu</p>
+        </MenuContainer>
+        <MainLayout>{
+          <p>main</p>
+        }</MainLayout>
+      </div>
     </BrowserRouter>
   </React.StrictMode>
 )
