+++ src/pages/dashboard/dashboard.style.tsx
@@ -1,16 +1,6 @@
 import styled from '@emotion/styled'
 
-export const DashboardContainer = styled('div')`
-    display: flex;
-    height: 100vh;
-    background-color: #beabab;
 
-`
-
-export const MenuContainer = styled('div')`
-    flex: 0 0 296px;
-    background-color: #ffffff;
-`
 
 export const ContentStyled = styled('div')`
   flex: 1;
