+++ src/pages/dashboard/layout/main/main.layout.tsx
@@ -1,3 +1,12 @@
-import styled from '@emotion/styled'
+import styled from '@emotion/styled';
 
-export const MainLayout = styled('div')``
+export const MainLayout = styled('div')`
+  display: flex;
+  height: 100vh;
+  background-color: var(--ui-color-background-light);
+`
+
+export const MenuContainer = styled('div')`
+  flex: 0 0 296px;
+  background-color: var(--ui-color-background-light);
+`
